cdk/requirements.txt
aws_cdk.aws_secretsmanager
aws_cdk.aws_cloudformation
aws_cdk.custom_resources
aws_cdk.aws_s3_notifications
aws_cdk.aws_s3
aws_cdk.aws_sns_subscriptions
aws_cdk.aws_lambda_event_sources
aws_cdk.aws_elasticsearch
aws_cdk.aws_cognito